+++ src/Can/HasOrientation.php
@@ -1,9 +1,10 @@
 <?php
 
 
-namespace AmpedWeb\GlideInABox\Traits;
+namespace AmpedWeb\GlideInABox\Can;
 
 use AmpedWeb\GlideInABox\Exceptions\InvalidOrientationException;
+use AmpedWeb\GlideInABox\Interfaces\Rotate;
 
 /**
  * Trait Orientation
@@ -11,9 +12,9 @@ use AmpedWeb\GlideInABox\Exceptions\InvalidOrientationException;
  * @property array $buildParams
  *
  * @see     https://glide.thephpleague.com/1.0/api/orientation/
- * @package AmpedWeb\GlideInABox\Traits
+ * @package AmpedWeb\GlideInABox\Can
  */
-trait Orientation
+trait HasOrientation
 {
     public static $ORIENTATION_AUTO = 'auto';
     public static $ORIENTATION_0 = '0';
@@ -33,22 +34,22 @@ trait Orientation
      *                            - '180'
      *                            - '270'.
      *                            Or you can use one of the convenient static variables:
-     *                            - Orientation::$ORIENTATION_AUTO
-     *                            - Orientation::$ORIENTATION_0
-     *                            - Orientation::$ORIENTATION_90
-     *                            - Orientation::$ORIENTATION_180
-     *                            - Orientation::$ORIENTATION_270
+     *                            - Rotate::AUTO
+     *                            - Rotate::R0
+     *                            - Rotate::R90
+     *                            - Rotate::R180
+     *                            - Rotate::R270
      *
      * @return $this
      * @throws InvalidOrientationException
      */
     public function orientation(string $orientation)
     {
-        if ($orientation !== static::$ORIENTATION_AUTO &&
-            $orientation !== static::$ORIENTATION_0 &&
-            $orientation !== static::$ORIENTATION_90 &&
-            $orientation !== static::$ORIENTATION_180 &&
-            $orientation !== static::$ORIENTATION_270
+        if ($orientation !== Rotate::AUTO &&
+            $orientation !== Rotate::R0 &&
+            $orientation !== Rotate::R90 &&
+            $orientation !== Rotate::R180 &&
+            $orientation !== Rotate::R270
         ) {
             throw new InvalidOrientationException();
         }