+++ src/Can/HasWatermarks.php
@@ -1,21 +1,24 @@
 <?php
 
 
-namespace AmpedWeb\GlideInABox\Traits;
+namespace AmpedWeb\GlideInABox\Can;
 
 use AmpedWeb\GlideInABox\Exceptions\InvalidDimensionException;
+use AmpedWeb\GlideInABox\Exceptions\InvalidFitException;
 use AmpedWeb\GlideInABox\Exceptions\InvalidMarkFitException;
 use AmpedWeb\GlideInABox\Exceptions\InvalidMarkPositionException;
+use AmpedWeb\GlideInABox\Interfaces\Fit;
+use AmpedWeb\GlideInABox\Interfaces\Position;
 
 /**
  * This trait provides watermarks functionality
  *
- * @package AmpedWeb\GlideInABox\Traits
+ * @package AmpedWeb\GlideInABox\Can
  * @link    https://glide.thephpleague.com/1.0/api/watermarks/
  */
-trait Watermarks
+trait HasWatermarks
 {
-    use DimensionParser;
+    use HasDimensionParser;
 
     /**
      * @property array $buildParams
@@ -55,7 +58,7 @@ trait Watermarks
      *
      * @param string $filename
      *
-     * @return Watermarks
+     * @return HasWatermarks
      */
     public function mark(string $filename)
     {
@@ -69,7 +72,7 @@ trait Watermarks
      *
      * @param string $dimension
      *
-     * @return Watermarks
+     * @return HasWatermarks
      * @throws InvalidDimensionException
      */
     public function markWidth(string $dimension)
@@ -84,7 +87,7 @@ trait Watermarks
      *
      * @param string $dimension
      *
-     * @return Watermarks
+     * @return HasWatermarks
      * @throws InvalidDimensionException
      */
     public function markHeight(string $dimension)
@@ -128,18 +131,18 @@ trait Watermarks
      *                    image data. The resulting image will match the width and height constraints without
      *                    distorting the image. See the crop page for more information.
      *
-     * @return Watermarks
+     * @return HasWatermarks
      * @throws InvalidMarkFitException
      */
     public function markFit(string $fit = 'contain')
     {
-        if ($fit !== Size::$FIT_CONTAIN &&
-            $fit !== Size::$FIT_CROP &&
-            $fit !== Size::$FIT_FILL &&
-            $fit !== Size::$FIT_MAX &&
-            $fit !== Size::$FIT_STRETCH
+        if ($fit !== Fit::CONTAIN &&
+            $fit !== Fit::CROP &&
+            $fit !== Fit::FILL &&
+            $fit !== Fit::MAX &&
+            $fit !== Fit::STRETCH
         ) {
-            throw new InvalidMarkFitException();
+            throw new InvalidFitException();
         }
 
         $this->buildParams['markfit'] = $fit;
@@ -154,7 +157,7 @@ trait Watermarks
      *
      * @param string $dimension
      *
-     * @return Watermarks
+     * @return HasWatermarks
      * @throws InvalidDimensionException
      */
     public function markX(string $dimension)
@@ -171,7 +174,7 @@ trait Watermarks
      *
      * @param string $dimension
      *
-     * @return Watermarks
+     * @return HasWatermarks
      * @throws InvalidDimensionException
      */
     public function markY(string $dimension)
@@ -189,7 +192,7 @@ trait Watermarks
      *
      * @param string $dimension
      *
-     * @return Watermarks
+     * @return HasWatermarks
      */
     public function markPad(string $dimension)
     {
@@ -215,19 +218,19 @@ trait Watermarks
      *                         - `bottom`,
      *                         - `bottom-right`.
      *
-     * @return Watermarks
+     * @return HasWatermarks
      */
     public function markPos(string $position = 'center')
     {
-        if ($position !== static::$MARKPOS_TOP_LEFT &&
-            $position !== static::$MARKPOS_TOP &&
-            $position !== static::$MARKPOS_TOP_RIGHT &&
-            $position !== static::$MARKPOS_LEFT &&
-            $position !== static::$MARKPOS_CENTER &&
-            $position !== static::$MARKPOS_RIGHT &&
-            $position !== static::$MARKPOS_BOTTOM_LEFT &&
-            $position !== static::$MARKPOS_BOTTOM &&
-            $position !== static::$MARKPOS_BOTTOM_RIGHT
+        if ($position !== Position::TOP_LEFT &&
+            $position !== Position::TOP &&
+            $position !== Position::TOP_RIGHT &&
+            $position !== Position::LEFT &&
+            $position !== Position::CENTER &&
+            $position !== Position::RIGHT &&
+            $position !== Position::BOTTOM_LEFT &&
+            $position !== Position::BOTTOM &&
+            $position !== Position::BOTTOM_RIGHT
         ) {
             throw new InvalidMarkPositionException();
         }