+++ src/Exceptions/InvalidFitException.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace AmpedWeb\GlideInABox\Exceptions;
+
+
+class InvalidFitException extends GlideUrlException
+{
+
+}