+++ src/Interfaces/Position.php
@@ -1,0 +1,18 @@
+<?php
+
+
+namespace AmpedWeb\GlideInABox\Interfaces;
+
+
+interface Position
+{
+    const TOP_LEFT = 'top-left';
+    const TOP = 'top';
+    const TOP_RIGHT = 'top-right';
+    const LEFT = 'left';
+    const RIGHT = 'right';
+    const CENTER = 'center';
+    const BOTTOM_LEFT = 'bottom-left';
+    const BOTTOM = 'bottom';
+    const BOTTOM_RIGHT = 'bottom-right';
+}