+++ tests/Feature/OrientationTest.php
@@ -6,7 +6,7 @@ namespace AmpedWeb\GlideInABox\Tests\Feature;
 
 use AmpedWeb\GlideInABox\Exceptions\InvalidOrientationException;
 use AmpedWeb\GlideInABox\Tests\ImageTestCase;
-use AmpedWeb\GlideInABox\Traits\Orientation;
+use AmpedWeb\GlideInABox\Can\HasOrientation;
 use AmpedWeb\GlideInABox\Util\GlideUrl;
 
 class OrientationTest extends ImageTestCase
@@ -31,37 +31,37 @@ class OrientationTest extends ImageTestCase
 
     public function testOrientationIsFluent()
     {
-        $glideUrl = $this->glideUrl->orientation(Orientation::$ORIENTATION_AUTO);
+        $glideUrl = $this->glideUrl->orientation(HasOrientation::$ORIENTATION_AUTO);
         $this->assertEquals($this->glideUrl, $glideUrl);
     }
 
     public function testOrientationSetsCorrectValueAuto()
     {
-        $this->glideUrl->orientation(Orientation::$ORIENTATION_AUTO);
+        $this->glideUrl->orientation(HasOrientation::$ORIENTATION_AUTO);
         $this->assertEquals('auto', $this->glideUrl->getParams()['or']);
     }
 
     public function testOrientationSetsCorrectValue0()
     {
-        $this->glideUrl->orientation(Orientation::$ORIENTATION_0);
+        $this->glideUrl->orientation(HasOrientation::$ORIENTATION_0);
         $this->assertEquals('0', $this->glideUrl->getParams()['or']);
     }
 
     public function testOrientationSetsCorrectValue90()
     {
-        $this->glideUrl->orientation(Orientation::$ORIENTATION_90);
+        $this->glideUrl->orientation(HasOrientation::$ORIENTATION_90);
         $this->assertEquals('90', $this->glideUrl->getParams()['or']);
     }
 
     public function testOrientationSetsCorrectValue180()
     {
-        $this->glideUrl->orientation(Orientation::$ORIENTATION_180);
+        $this->glideUrl->orientation(HasOrientation::$ORIENTATION_180);
         $this->assertEquals('180', $this->glideUrl->getParams()['or']);
     }
 
     public function testOrientationSetsCorrectValue270()
     {
-        $this->glideUrl->orientation(Orientation::$ORIENTATION_270);
+        $this->glideUrl->orientation(HasOrientation::$ORIENTATION_270);
         $this->assertEquals('270', $this->glideUrl->getParams()['or']);
     }
 }